requirements.txt
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
mlflow>=1.20.0
joblib>=1.0.0
ipykernel>=6.0.0  # For notebook support
pytest>=6.0.0  # For testing
pyyaml==6.0.2
xgboost==3.0.0